changed coordinate and size system to use float. Coords are always tracked by float, converted to int only on draw to choose a correct pixel to draw to.

+++ include/Entity.h
@@ -5,10 +5,10 @@
 
 class Entity {
 public:
-    int x;
-    int y;
-    int w;
-    int h;
+    float x;
+    float y;
+    float w;
+    float h;
     float* td;
 
     // Functions:
